# Conflicts:
#	lib/src/main/java/fr/onema/lib/file/FileManager.java
#	lib/src/test/java/fr/onema/lib/File/FileManagerTest.java

+++ lib/src/main/java/fr/onema/lib/file/Parser.java
@@ -4,9 +4,15 @@ package fr.onema.lib.file;
 import fr.onema.lib.virtualizer.entry.ReferenceEntry;
 import fr.onema.lib.virtualizer.entry.VirtualizerEntry;
 
-// TODO : complete -> replace into factory
+/***
+ * Classe utilitaire permettant de transformer une ligne CSV en entités
+ */
 public class Parser {
-    // TODO : complete
+    /***
+     * Parse la ligne et la transforme en entrée de référence
+     * @param line La ligne CSV à parser
+     * @return L'objet {@link ReferenceEntry} associé
+     */
     public static ReferenceEntry parseReference(String line) {
         String[] s = line.split(",");
         if (s.length != 6) {
@@ -20,10 +26,14 @@ public class Parser {
                 Short.parseShort(s[5]));
     }
 
-    // TODO : complete
+    /***
+     * Parse la ligne et la transforme en entrée virtualisée
+     * @param line La ligne CSV à parser
+     * @return L'objet {@link VirtualizerEntry} associé
+     */
     public static VirtualizerEntry parseVirtualizer(String line) {
         String[] s = line.split(",");
-        if(s.length != 15) {
+        if (s.length != 15) {
             throw new IllegalArgumentException();
         }
         return new VirtualizerEntry(Long.parseLong(s[0]),
@@ -33,9 +43,9 @@ public class Parser {
                 Short.parseShort(s[4]),
                 Short.parseShort(s[5]),
                 Short.parseShort(s[6]),
-                Short.parseShort(s[7]),
-                Short.parseShort(s[8]),
-                Short.parseShort(s[9]),
+                Double.parseDouble(s[7]),
+                Double.parseDouble(s[8]),
+                Double.parseDouble(s[9]),
                 Short.parseShort(s[10]),
                 Short.parseShort(s[11]),
                 Short.parseShort(s[12]),
